+++ bindings/storage.deps
@@ -27,6 +27,7 @@ DEPENDENCIES =							\
 	${top_srcdir}/storage/Filesystems/Iso9660.h		\
 	${top_srcdir}/storage/Filesystems/Udf.h			\
 	${top_srcdir}/storage/Filesystems/Nfs.h 		\
+	${top_srcdir}/storage/Filesystems/Mountable.h		\
 	${top_srcdir}/storage/Filesystems/Filesystem.h		\
 	${top_srcdir}/storage/Filesystems/BlkFilesystem.h	\
 	${top_srcdir}/storage/Devices/Gpt.h			\
