+++ bindings/storage.i
@@ -36,6 +36,7 @@
 #include "storage/UsedFeatures.h"
 
 #include "storage/Devices/Device.h"
+#include "storage/Filesystems/Mountable.h"
 #include "storage/Filesystems/Filesystem.h"
 #include "storage/Filesystems/BlkFilesystem.h"
 #include "storage/Filesystems/Ext.h"
@@ -98,6 +99,7 @@
 %include "../../storage/UsedFeatures.h"
 
 %include "../../storage/Devices/Device.h"
+%include "../../storage/Filesystems/Mountable.h"
 %include "../../storage/Filesystems/Filesystem.h"
 %include "../../storage/Filesystems/BlkFilesystem.h"
 %include "../../storage/Filesystems/Ext.h"
